Created a Database to Store Player Information (Name, Life, Poison)
Created a database that will hold all of the information from all
players so that this data doesn't need to be sent through the intents
after every activity transition. DBManager is the database manager that
handles the SQL queries and the Player class is used to hold individual
player information.